push for screenshot

+++ app/api/files/[fileId]/route.ts
@@ -1,5 +1,6 @@
 import { currentProfile } from "@/actions/currentProfile";
 import { db } from "@/lib/db";
+import { MemberRole } from "@prisma/client";
 import { NextResponse } from "next/server";
 import { UTApi } from "uploadthing/server"
 
@@ -27,6 +28,18 @@ export async function DELETE(req: Request, { params }: { params: { fileId: strin
 
         //TODO: check if the user that wants to delete the file is an admin
 
+        const member = await db.member.findFirst({
+            where: {
+                profileId: profile.id,
+                projectId: file.projectID
+            }
+        })
+
+        if (member?.role != MemberRole.ADMIN) {
+            return new NextResponse("Only admins can create tasks", { status: 401 })
+        }
+
+        //TODO:----------------------------------------------------------------
         const utapi = new UTApi();
         const newUrl = file.fileUrl.substring(file.fileUrl.lastIndexOf("/") + 1);
         await utapi.deleteFiles(newUrl)
